Imagem3

+++ src/mundo-veg/Controllers/UsuarioPjsController.cs
@@ -15,66 +15,12 @@ namespace mundo_veg.Controllers
     public class UsuarioPjsController : Controller
     {
         private readonly ApplicationDbContext _context;
-        private string _filePath;
-        public UsuarioPjsController(ApplicationDbContext context, IWebHostEnvironment env)
+        public UsuarioPjsController(ApplicationDbContext context)
         {
-            _filePath = env.WebRootPath;
             _context = context;
         }
 
-        public IActionResult Login()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Login([Bind("Email,Senha")] UsuarioPj usuario)
-        {
-            var user = await _context.UsuarioPjs
-                .FirstOrDefaultAsync(m => m.Email == usuario.Email);
-
-            if (user == null)
-            {
-                ViewBag.Message = "Usuário e/ou senha inválidos!";
-                return View();
-            }
-
-            bool isSenhaOk = BCrypt.Net.BCrypt.Verify(usuario.Senha, user.Senha);
-
-            if (isSenhaOk)
-            {
-                var claims = new List<Claim>
-                {
-                    new Claim(ClaimTypes.Name, user.Nome),
-                    new Claim(ClaimTypes.NameIdentifier, user.Nome),
-                };
-
-                var userPjIdentity = new ClaimsIdentity(claims, "login");
-
-                ClaimsPrincipal principal = new ClaimsPrincipal(userPjIdentity);
-
-                var props = new AuthenticationProperties
-                {
-                    AllowRefresh = true,
-                    ExpiresUtc = DateTime.Now.ToLocalTime().AddDays(7),
-                    IsPersistent = true
-                };
-
-                await HttpContext.SignInAsync(principal, props);
-
-                return Redirect("/");
-            }
-
-            ViewBag.Message = "Usuário e/ou Senha inválidos!";
-            return View();
-        }
-
-        [AllowAnonymous]
-        public async Task<IActionResult> Logout()
-        {
-            await HttpContext.SignOutAsync();
-            return RedirectToAction("Login", "UsuarioPjs");
-        }
+     
 
         [AllowAnonymous]
         public IActionResult AccessDenied()
@@ -128,14 +74,26 @@ namespace mundo_veg.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,CNPJ,Nome,Email,Telefone,Rua,Numero,Bairro,Cidade,Estado,CEP,Horario_funcionamento,Faz_entrega,Tempo_entrega,Descricao,Contato,Senha,Imagem")] UsuarioPj usuarioPj, IFormFile anexo)
+        public async Task<IActionResult> Create([Bind("Id,CNPJ,Nome,Email,Telefone,Rua,Numero,Bairro,Cidade,Estado,CEP,Horario_funcionamento,Faz_entrega,Tempo_entrega,Descricao,Contato,Senha")] UsuarioPj usuarioPj)
         {
             if (ModelState.IsValid)
             {
-                if (!ValidaImagem(anexo))
+                var userPj = await _context.UsuarioPjs
+                   .FirstOrDefaultAsync(m => m.Email == usuarioPj.Email);
+                if (userPj != null)
+                {
+                    ViewBag.Message = "Esse email já foi cadastrado!";
                     return View(usuarioPj);
-                var nomeImg = SalvarArquivo(anexo);
-                usuarioPj.Imagem = nomeImg;
+                }
+                else
+                {
+                 var userPf = await _context.UsuarioPfs.FirstOrDefaultAsync(m=>m.Email == usuarioPj.Email);
+                    if(userPf != null)
+                    {
+                        ViewBag.Message = "Esse email já foi cadastrado!";
+                        return View(usuarioPj);
+                    }
+                }
                 
                 usuarioPj.Senha = BCrypt.Net.BCrypt.HashPassword(usuarioPj.Senha);
                 _context.Add(usuarioPj);
@@ -144,42 +102,7 @@ namespace mundo_veg.Controllers
             }
             return View(usuarioPj);
         }
-        public bool ValidaImagem(IFormFile anexo)
-        {
-            switch (anexo.ContentType)
-            {
-                case "image/jpg":
-                    return true;
-                case "image/jpeg":
-                    return true;
-                case "image/bmp":
-                    return true;
-                case "image/png":
-                    return true;
-                default:
-                    return false;
-                    break;
 
-            }
-        }
-        public string SalvarArquivo(IFormFile anexo)
-        {
-            var nome = Guid.NewGuid().ToString() + anexo.FileName;
-
-            var filePath = _filePath + "\\imagens";
-
-            if (!Directory.Exists(filePath)) ;
-            {
-                Directory.CreateDirectory(filePath);
-            }
-
-            using (var stream = System.IO.File.Create(filePath + "\\" + nome))
-            {
-                anexo.CopyToAsync(stream);
-            }
-            return nome;
-
-        }
         // GET: UsuarioPjs/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -201,7 +124,7 @@ namespace mundo_veg.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-         public async Task<IActionResult> Edit(int id, [Bind("Id,CNPJ,Nome,Email,Telefone,Rua,Numero,Bairro,Cidade,Estado,CEP,Horario_funcionamento,Faz_entrega,Tempo_entrega,Descricao,Contato,Senha,Imagem")] UsuarioPj usuarioPj, IFormFile anexo)
+         public async Task<IActionResult> Edit(int id, [Bind("Id,CNPJ,Nome,Email,Telefone,Rua,Numero,Bairro,Cidade,Estado,CEP,Horario_funcionamento,Faz_entrega,Tempo_entrega,Descricao,Contato,Senha")] UsuarioPj usuarioPj, IFormFile anexo)
         {
             if (id != usuarioPj.Id)
             {
@@ -213,11 +136,6 @@ namespace mundo_veg.Controllers
                 try
                 {
                     usuarioPj.Senha = BCrypt.Net.BCrypt.HashPassword(usuarioPj.Senha);
-
-                    if (!ValidaImagemEdit(anexo))
-                        return View(usuarioPj);
-                    var nomeImg = SalvarArquivo(anexo);
-                    usuarioPj.Imagem = nomeImg;
 
                     _context.Update(usuarioPj);
                     await _context.SaveChangesAsync();
@@ -236,41 +154,6 @@ namespace mundo_veg.Controllers
                 return RedirectToAction("Details", "UsuarioPjs", new { id = usuarioPj.Id });
             }
             return View(usuarioPj);
-        }
-        public bool ValidaImagemEdit(IFormFile anexo)
-        {
-            switch (anexo.ContentType)
-            {
-                case "image/jpg":
-                    return true;
-                case "image/jpeg":
-                    return true;
-                case "image/bmp":
-                    return true;
-                case "image/png":
-                    return true;
-                default:
-                    return false;
-                    break;
-
-            }
-        }
-        public string SalvarArquivoEdit(IFormFile anexo)
-        {
-            var nome = Guid.NewGuid().ToString() + anexo.FileName;
-
-            var filePath = _filePath + "\\imagens";
-
-            if (!Directory.Exists(filePath)) ;
-            {
-                Directory.CreateDirectory(filePath);
-            }
-
-            using (var stream = System.IO.File.Create(filePath + "\\" + nome))
-            {
-                anexo.CopyToAsync(stream);
-            }
-            return nome;
 
         }
         // GET: UsuarioPjs/Delete/5
@@ -301,15 +184,15 @@ namespace mundo_veg.Controllers
                 return Problem("Entity set 'ApplicationDbContext.UsuarioPjs'  is null.");
             }
             var usuarioPj = await _context.UsuarioPjs.FindAsync(id);
-            string filePathName = _filePath + "\\imagens\\" + usuarioPj.Imagem;
 
-            if (System.IO.File.Exists(filePathName))
-                System.IO.File.Delete(filePathName);
-            _context.UsuarioPjs.Remove(usuarioPj);
-            
+            if (usuarioPj != null)
+            {
+                _context.UsuarioPjs.Remove(usuarioPj);
+            }
 
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            await HttpContext.SignOutAsync();
+            return RedirectToAction("Index");
         }
 
         private bool UsuarioPjExists(int id)
